mimic new file structure in unit tests
